Update README.md
This commit includes changes to the README file, which have been generated with the help of the GitHub Profile README Generator available at https://rahuldkjain.github.io/gh-profile-readme-generator/. The generator allowed for the easy creation and customization of our GitHub profile README, enhancing the overall presentation of our project or profile information.

The README content has been modified and improved to provide a more informative and visually appealing representation of the project or profile. This tool streamlined the process of creating a well-structured README, making it easier for users to understand our project's purpose and how to use it effectively.